Add PR_{GET,SET}_NO_NEW_PRIVS to prevent execve from granting privs
With this change, calling prctl(PR_SET_NO_NEW_PRIVS, 1, 0, 0, 0)
disables privilege granting operations at execve-time.  For example,
with this but set, a process will not be able to execute a setuid binary
to change their uid or gid.  The same is true for file capabilities.

Additionally, LSM_UNSAFE_NO_NEW_PRIVS is defined to ensure that
LSMs respect the requested behavior.

This functionality is desired for the proposed seccomp filter patch
series.  By using PR_SET_NO_NEW_PRIVS, it will be possible for a task
to modify the system call behavior for itself and its child tasks
without ever being able to impact the behavior of a more privileged
task.

In the future, this may also allow other operations, like chroot, which
require privileges to be provided to unprivileged users if the caller
can never affect a privileged task.

Note, this patch causes execve to fail when PR_SET_NO_NEW_PRIVS is
set and AppArmor is in use.  It is fixed in a subsequent patch.

Signed-off-by: Will Drewry <wad@chromium.org>
Signed-off-by: Andy Lutomirski <luto@amacapital.net>
Acked-by: Eric Paris <eparis@redhat.com>