sk_run_filter: add support for custom load_pointer
This change allows CONFIG_SECCOMP to make use of BPF programs for
user-controlled system call filtering (as shown in this patch series).

To minimize the impact on existing BPF evaluation, function pointer
use must be declared at sk_chk_filter-time.  This allows ancillary
load instructions to be generated that use the function pointer rather
than adding _any_ code to the existing LD_* instruction paths.

Crude performance numbers using udpflood -l 10000000 against dummy0.
3 trials just udpflooding, 3 with tcpdump. Averaged then differenced.

[v11 #s]
* x86 32-bit (Atom N570 @ 1.66 GHz 2 core HT) [stackprot]:
          (tcpdump)   (no bpf)
- Without:  88.19s  -  70.98s  =  17.21s
- With:     90.97s  -  74.96s  =  15.98s
- Slowdown per packet: -123 nanoseconds

* x86 64-bit (Atom N570 @ 1.66 GHz 2 core HT) [stackprot]:
          (tcpdump)   (no bpf)
- Without: 117.80s  - 103.99s  =  13.80s
- With:    116.81s  - 101.85s  =  14.96s
- Slowdown per packet: 116 nanoseconds

In this round of benchmarking, the gap seems narrower and I've attempted
to take extra steps to ensure a quiescent system (especially given the
wild variability I was seeing on 64-bit last time).  It's not clear to
me if this overhead (on an Atom-class system) is acceptable or not.
Irrespective, I had planned to follow on to this patch set updating
every in-kernel callsite of sk_run/chk_filter to use bpf_*_filter
directly and alleviate the bounce through. I was hoping that could be
done independently, however.  If it would make sense to do that first
or just duplicate the sk_run_filter code in seccomp.c initially, please
let me know.

v13: rebase on to 88ebdda6159ffc15699f204c33feb3e431bf9bdc
v12: rebase to linux-next
v11: - un-inline sk_*_filter to adhere to the existing export symbol use
       and not break bpf_jit_free (markus@chromium.org)
     - updated benchmarks
v10: - converted length to a u32 on the struct because it is passed in
       at bpf_run_filter time anyway. (indan@nul.nu)
     - added a comment about the LD_*_ falling through (joe@perches.com)
     - more clean up (pointer->load, drop SKB macro, comments) (indan@nul.nu)
v9: - n/a
v8: - fixed variable positioning and bad cast (eric.dumazet@gmail.com)
    - no longer passes A as a pointer (inspection of x86 asm shows A is
      %ebx again; thanks eric.dumazet@gmail.com)
    - cleaned up switch macros and expanded use
      (joe@perches.com, indan@nul.nu)
    - added length fn pointer and handled LD_W_LEN/LDX_W_LEN
    - moved from a wrapping struct to a typedef for the function
      pointer. (matches existing function pointer style)
    - added comprehensive comment above the typedef.
    - benchmarks
v7: - first cut

Signed-off-by: Will Drewry <wad@chromium.org>
Acked-by: Eric Dumazet <eric.dumazet@gmail.com>